Use own build stages for librespot and shairport-sync; add airplay-2 build support; disable NGINX by default

+++ setup.sh
@@ -1,9 +1,13 @@
 #!/bin/sh
 set -e
 
-# Prepare dbus environment
+# Prepare dbus-daemon environment
 dbus-uuidgen --ensure
 
+
+#
+# SETUP SNAPCAST
+#
 
 # SNAPCAST: Create default configuration for snapserver
 SNAPCAST_CONFIG=""
@@ -12,11 +16,19 @@ if [ "${PIPE_CONFIG_ENABLED}" -eq 1 ]; then
     SNAPCAST_CONFIG="${SNAPCAST_CONFIG}source = pipe://${PIPE_PATH}?name=${PIPE_SOURCE_NAME}&mode=${PIPE_MODE}\n"
 fi
 if [ "${AIRPLAY_CONFIG_ENABLED}" -eq 1 ]; then
-    SNAPCAST_CONFIG="${SNAPCAST_CONFIG}source = airplay:///shairport-sync?name=${AIRPLAY_SOURCE_NAME}&port=5000\n"
+    if [ "${BUILD_AIRPLAY_VERSION}" -eq 2 ]; then
+        AIRPLAY_PORT="7000"
+        echo "[SETUP] Configuring Snapserver for Airplay 2..."
+    else
+        AIRPLAY_PORT="5000"
+        echo "[SETUP]  Configuring Snapserver for Airplay 2..."
+    if
+
+    SNAPCAST_CONFIG="${SNAPCAST_CONFIG}source = airplay:///shairport-sync?name=${AIRPLAY_SOURCE_NAME}&port=${AIRPLAY_PORT}\n"
 fi
 if [ "${SPOTIFY_CONFIG_ENABLED}" -eq 1 ]; then
     if [ -z "${SPOTIFY_USERNAME}" ] || [ -z "${SPOTIFY_PASSWORD}" ]; then
-        echo "Error: Cannot create spotify configuration! Username and/or password are not set!"
+        echo "[SETUP]  Error: Cannot create spotify configuration! Username and/or password are not set!"
     else
         SNAPCAST_CONFIG="${SNAPCAST_CONFIG}source = spotify:///librespot?name=${SPOTIFY_SOURCE_NAME}&username=${SPOTIFY_USERNAME}&password=${SPOTIFY_PASSWORD}&devicename=${SPOTIFY_DEVICE_NAME}&bitrate=${SPOTIFY_BITRATE}\n"
     fi
@@ -38,21 +50,56 @@ cp -n /tmp/snapserver.conf /app/config/snapserver.conf
 rm /tmp/snapserver.conf
 
 
-# Prepare NGINX configuration
-# NGINX: generate self-signed ssl certificates, if no certs are existant
-if [ -s /app/certs/snapserver.pem ] || [ -s /app/certs/snapserver.key ] || [ "${NGINX_SKIP_CERT_GENERATION}" -eq 1 ]; then
-    echo "Server SSL certificates for NGINX already exist, skipping generation."
-else
-    echo "Generating self-signed certificates for NGINX."
-    openssl req -x509 -nodes -days 365 -newkey rsa:2048 \
-        -keyout /app/certs/snapserver.key -out /app/certs/snapserver.pem \
-        -subj "/C=DE/ST=Bavaria/L=Nuremberg/O=Snapserver/CN=snapserver"
+#
+# SETUP NGINX
+#
+if [ "${NGINX_ENABLED}" -eq 1 ]; then
+    # NGINX: generate self-signed ssl certificates, if no certs are existant
+    if [ -s /app/certs/snapserver.pem ] || [ -s /app/certs/snapserver.key ] || [ "${NGINX_SKIP_CERT_GENERATION}" -eq 1 ]; then
+        echo "[SETUP] Server SSL certificates for NGINX already exist, skipping generation."
+    else
+        echo "[SETUP] Generating self-signed certificates for NGINX..."
+        openssl req -x509 -nodes -days 365 -newkey rsa:2048 \
+            -keyout /app/certs/snapserver.key -out /app/certs/snapserver.pem \
+            -subj "/C=DE/ST=Bavaria/L=Nuremberg/O=Snapserver/CN=snapserver"
+    fi
+
+    # NGINX: Replace port in NGINX config
+    # Note: sed cannot directly replace in-file as it cannot create a temporary file under this directory.
+    sed "s/443/${NGINX_HTTPS_PORT}/g" /etc/nginx/http.d/default.conf > /tmp/default.conf
+    cat /tmp/default.conf > /etc/nginx/http.d/default.conf
+    rm /tmp/default.conf
+
+    # NGINX: Create supervisord configuration
+    NGINX_SUPERVISORD_CONFIG="
+    [program:nginx]\n
+    command=/usr/sbin/nginx -g 'daemon off;'\n
+    autostart=true\n
+    autorestart=true\n
+    startsecs=3\n
+    startretries=5\n
+    priority=40\n
+    "
+    echo -e "${NGINX_SUPERVISORD_CONFIG}" > /app/supervisord/nginx.ini
 fi
 
-# NGINX: Replace port in NGINX config
-# Note: sed cannot directly replace in-file as it cannot create a temporary file under this directory.
-sed "s/443/${NGINX_HTTPS_PORT}/g" /etc/nginx/http.d/default.conf > /tmp/default.conf
-cat /tmp/default.conf > /etc/nginx/http.d/default.conf
-rm /tmp/default.conf
+
+#
+# SETUP SHAIRPORT-SYNC
+#
+
+# Prepare Shairport-Sync Airplay-2 configuration
+if [ "${BUILD_AIRPLAY_VERSION}" -eq 2 ]; then
+    NQPTP_SUPERVISORD_CONFIG="
+    [program:nqptp]\n
+    command=/usr/local/bin/nqptp\n
+    autostart=true\n
+    autorestart=true\n
+    startsecs=3\n
+    startretries=5\n
+    priority=21\n
+    "
+    echo -e "${NQPTP_SUPERVISORD_CONFIG}" > /app/supervisord/nqptp.ini
+fi
 
 exit 0
